refactor: migrate from Vite to TSDown build system
- Remove Vite configuration and dependencies
- Add TSDown configuration and build scripts
- Update project structure for TSDown compatibility
- Update documentation and package.json
