+++ README.md
@@ -93,7 +93,7 @@ The generated class names will look like:
 - SomeComponent\_\_wrapper_abcd1_1
 - SomeComponent\_\_container_abcd2_5
 
-> [!WARNING]
+> [!IMPORTANT]
 > Please note that the `lineNumber` option mirrors the behavior of Vite's default class name generation when using preprocessors like Sass, Less, or Stylus. The line number is calculated based on the compiled CSS, where empty lines between selectors and comments are typically removed. This can lead to discrepancies between the line numbers in the source files and the compiled output, potentially resulting in inaccurate line numbers in the generated class names.
 >
 > Additionally, in Vue files, the line count always starts from the `<style module>` tag, regardless of where it is placed within the file. This means that the line numbers in generated class names will be relative to the position of the `<style module>` tag, not the beginning of the file.
