+++ eslint.config.js
@@ -1,0 +1,3 @@
+import { sxzz } from '@sxzz/eslint-config'
+
+export default sxzz()
