+++ package.json
@@ -3,9 +3,23 @@
   "version": "1.3.2",
   "description": "Adds the filename without the -module suffix to the class names of CSS modules.",
   "type": "module",
-  "main": "dist/index.cjs",
-  "module": "dist/index.esm.mjs",
-  "types": "dist/index.d.ts",
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": {
+        "import": "./dist/index.d.ts",
+        "require": "./dist/index.d.cts"
+      },
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    },
+    "./package.json": "./package.json"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
   "keywords": [
     "vite",
     "vite-plugin",
@@ -44,7 +58,11 @@
     "LICENSE"
   ],
   "scripts": {
-    "build": "vite build",
+    "lint": "eslint --cache .",
+    "lint:fix": "npm run lint --fix",
+    "build": "tsdown",
+    "dev": "tsdown --watch",
+    "format": "prettier --cache --write .",
     "prepublishOnly": "npm run build",
     "release": "standard-version",
     "release:major": "npm run release -- --release-as major",
@@ -64,13 +82,18 @@
     "prepare": "husky"
   },
   "devDependencies": {
+    "@sxzz/eslint-config": "^6.1.2",
+    "@sxzz/prettier-config": "^2.2.1",
     "@types/node": "^20.12.7",
-    "@vitest/coverage-v8": "^3.0.5",
+    "@vitest/coverage-v8": "^3.1.2",
+    "eslint": "^9.25.0",
     "husky": "^9.1.7",
+    "prettier": "^3.5.3",
     "standard-version": "^9.5.0",
-    "typescript": "^5.2.2",
-    "vite": "^5.2.10",
-    "vite-plugin-dts": "^3.9.0",
-    "vitest": "^3.0.5"
-  }
+    "tsdown": "^0.9.1",
+    "typescript": "^5.8.3",
+    "vite": "^6.3.2",
+    "vitest": "^3.1.2"
+  },
+  "prettier": "@sxzz/prettier-config"
 }
